refactor: 컨트롤러, 서비스, 레파지토리 구조 도입
